named_entity_recognition/requirements.txt
grpcio==1.14.1
grpcio-tools==1.14.1
nltk
scikit-learn
numpy
scipy
